2010-11-09 15:13 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * contrib/hbplist     + Enabled hbzebra.
  * contrib/hbzebra/hbzebra.hbp
    + Added optional cairo dependency.
    + Enabled cairo dependent components if cairo is detected.
    - Deleted hard-coded reference to cairo lib.
    + Adjusted reference to hbcairo.hbc.

  * contrib/hbzebra/itf.c
  * contrib/hbzebra/codabar.c
    * Silenced three warnings.
    ; TOFIX: Remains one: 
             ../../../../../contrib/hbzebra/code128.c: In function 'hb_zebra_create_code128':
             ../../../../../contrib/hbzebra/code128.c:316:19: warning: suggest parentheses around '&&' within '||'

  * contrib/hbzebra/core.c
    ! Replaced hbcairo.h header reference with hbstack.h

  * contrib/hbzebra/cairo.c
  * contrib/hbzebra/hbzebra.hbc
    ! Added EOL at EOF.
    * Indenting.

+++ harbour/contrib/hbzebra/codabar.c
@@ -88,7 +88,7 @@ static char _codabar_charno( char ch )
       char * ptr = strchr( s_symbols, ch );
       if( ptr && *ptr )
          return ptr - s_symbols + 10;
-   } 
+   }
    return -1;
 }
 
@@ -153,8 +153,8 @@ PHB_ZEBRA hb_zebra_create_codabar( const char * szCode, HB_SIZE nLen, int iFlags
 
    pZebra->pBits = hb_bitbuffer_create();
 
-   if( iLen == 0 || _codabar_charno( szCode[ 0 ] ) < 16 )
-      _codabar_add( pZebra->pBits, s_code[ _codabar_charno( 'A' ) ], iFlags, HB_FALSE );  /* Default start A */
+   if( iLen == 0 || _codabar_charno( ( int ) szCode[ 0 ] ) < 16 )
+      _codabar_add( pZebra->pBits, s_code[ ( int ) _codabar_charno( 'A' ) ], iFlags, HB_FALSE );  /* Default start A */
 
    for( i = 0; i < iLen; i++ )
    {
@@ -163,7 +163,7 @@ PHB_ZEBRA hb_zebra_create_codabar( const char * szCode, HB_SIZE nLen, int iFlags
    }
 
    if( iLen == 0 || _codabar_charno( szCode[ i - 1 ] ) < 16 )
-      _codabar_add( pZebra->pBits, s_code[ _codabar_charno( 'B' ) ], iFlags, HB_TRUE );  /* Default stop B */
+      _codabar_add( pZebra->pBits, s_code[ ( int ) _codabar_charno( 'B' ) ], iFlags, HB_TRUE );  /* Default stop B */
 
    return pZebra;
 }
